Refactor InstIntepreterHelpers into several smaller files grouped by function
